[REFACTOR]: Codebase refactor
