+++ src/environments/environments.js
@@ -1,6 +1,6 @@
 export const env = {
 	argoUrl: "http://localhost:8080",
 	argoToken:
-		"eyJhbGciOiJSUzI1NiIsImtpZCI6ImJkZUdfNlZiOVBXLUhUY0tNNXlQblQ0WHlLeW9pSGY5eGlsTUtNNzhDQ28ifQ.eyJhdWQiOlsiaHR0cHM6Ly9rdWJlcm5ldGVzLmRlZmF1bHQuc3ZjLmNsdXN0ZXIubG9jYWwiXSwiZXhwIjoxNzgyMjkyOTQ3LCJpYXQiOjE3NTA3NTY5NDcsImlzcyI6Imh0dHBzOi8va3ViZXJuZXRlcy5kZWZhdWx0LnN2Yy5jbHVzdGVyLmxvY2FsIiwianRpIjoiOGMzZDYxZGUtNWIxYS00ZjBiLWIzOTktYzY5OGRlMjk1NjcyIiwia3ViZXJuZXRlcy5pbyI6eyJuYW1lc3BhY2UiOiJhcmdvIiwibm9kZSI6eyJuYW1lIjoibWluaWt1YmUiLCJ1aWQiOiI1ZTlmZmZkMC04MzY0LTRkMWMtYmJjMS1mMWJlZGM2M2U4N2YifSwicG9kIjp7Im5hbWUiOiJhcmdvLXNlcnZlci03OTZmZDU4ZmM3LWg1eGpsIiwidWlkIjoiYmYzNmE4ODYtMWMzMS00YThhLWIwMzMtZmY1YzA4NmUyNjRjIn0sInNlcnZpY2VhY2NvdW50Ijp7Im5hbWUiOiJhcmdvLXNlcnZlciIsInVpZCI6Ijc3ZDg5NjhiLTQzM2ItNGQ5YS04NDBkLTY5YjVkYjNkOTEzMCJ9LCJ3YXJuYWZ0ZXIiOjE3NTA3NjA1NTR9LCJuYmYiOjE3NTA3NTY5NDcsInN1YiI6InN5c3RlbTpzZXJ2aWNlYWNjb3VudDphcmdvOmFyZ28tc2VydmVyIn0.UnRtqFBIXaBIsDMtFFY1B1ie0Kw75e4SGMrQ2mVpjySSK0syDt7YcjHZOoMku-zL2ZHYx44PvKxsEcydiUbMhbY0mboQXhFXWnUuS1bQa8v2U6a8uWFDcjP5b-888AaVAcUHlf1K_VnwcA-nK9EbwsvGDyf3gbzIAkwzD_6AKiL8_7NpWIsoYWz4fx8FTIGXV9AC9FG0TZu_IXIALYbAmE1PWrIdJ7zVcD5aX-ft1GT7DArqpcFuHMsubIhDwqsFoiahA-vXBcYhHbgkubISbvdIfg9q-7DhhJQN6l6w75PUulhO6-L3bFg9mwaIO97BEssu4CHq2x_s-QFW5GgvIA",
+		"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6ImJkZUdfNlZiOVBXLUhUY0tNNXlQblQ0WHlLeW9pSGY5eGlsTUtNNzhDQ28ifQ.eyJhdWQiOlsiaHR0cHM6Ly9rdWJlcm5ldGVzLmRlZmF1bHQuc3ZjLmNsdXN0ZXIubG9jYWwiXSwiZXhwIjoxNzgyODE1MjcyLCJpYXQiOjE3NTEyNzkyNzIsImlzcyI6Imh0dHBzOi8va3ViZXJuZXRlcy5kZWZhdWx0LnN2Yy5jbHVzdGVyLmxvY2FsIiwianRpIjoiNTYzZWY3OTctOTdjNy00NzhkLWFhZDItODAyZGNmMjVhY2JmIiwia3ViZXJuZXRlcy5pbyI6eyJuYW1lc3BhY2UiOiJhcmdvIiwibm9kZSI6eyJuYW1lIjoibWluaWt1YmUiLCJ1aWQiOiI1ZTlmZmZkMC04MzY0LTRkMWMtYmJjMS1mMWJlZGM2M2U4N2YifSwicG9kIjp7Im5hbWUiOiJhcmdvLXNlcnZlci03OTZmZDU4ZmM3LWg1eGpsIiwidWlkIjoiYmYzNmE4ODYtMWMzMS00YThhLWIwMzMtZmY1YzA4NmUyNjRjIn0sInNlcnZpY2VhY2NvdW50Ijp7Im5hbWUiOiJhcmdvLXNlcnZlciIsInVpZCI6Ijc3ZDg5NjhiLTQzM2ItNGQ5YS04NDBkLTY5YjVkYjNkOTEzMCJ9LCJ3YXJuYWZ0ZXIiOjE3NTEyODI4Nzl9LCJuYmYiOjE3NTEyNzkyNzIsInN1YiI6InN5c3RlbTpzZXJ2aWNlYWNjb3VudDphcmdvOmFyZ28tc2VydmVyIn0.LQGXqPO5FN5tD0kGjthlx3wGN9gSVKhy4EjW2BywwZGt-9eF_PoxEvPBxXlzgUtgj3vJr9dUHxxkxK_7N5YvAlUM12_njaiLGXSaoi0vuCjMI53bUf08wWvqo2jPbvNqtc0udyd4_UHMnKxA43BSJQ76s01AGSAL7URYeeKpRJyrUCzfaCzbOVBkWMAAYszqwIO1sAOX73-Am-oJ19rjHOSp7QlkntVWYYqQ9ZPBKTrXC8Noi3ouBvqgZPWPQUFyD2B3J1NBFPPiAS_OPUwJClwgwZiKQaNPNp-DmGhkFKQErmNMwWHXfHu--c5KgCKyJZmayMLatIGGQQtin5ojkw",
 	sdkImage: "amarthya03/icmr-fhir-backend-image",
 };
