+++ src/react.env
@@ -1,3 +1,3 @@
 REACT_APP_ARGO_SERVER_URL=http://localhost:8080
 REACT_APP_ARGO_TOKEN=your_generated_token_here       
-REACT_APP_CONTAINER_IMAGE=suryansh72/icmr-fhir-cli:latest 
+REACT_APP_CONTAINER_IMAGE=env.sdkImage
