+++ src/styles.module.css
@@ -1,0 +1,214 @@
+.appContainer {
+	display: flex;
+	width: 100vw;
+	height: 100vh;
+	font-family: "Inter", sans-serif;
+}
+
+.header {
+	position: absolute;
+	top: 0;
+	left: 0;
+	right: 0;
+	height: 60px;
+	background-color: #ffffff;
+	border-bottom: 1px solid #e0e0e0;
+	display: flex;
+	align-items: center;
+	justify-content: space-between;
+	padding: 0 20px;
+	z-index: 900;
+	box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+}
+
+.headerTitle {
+	font-size: 1.4rem;
+	font-weight: 600;
+	margin: 0;
+	color: #333;
+}
+
+.headerPlaceholder {
+	display: flex;
+	gap: 15px;
+}
+
+.sidebar {
+	width: 240px;
+	padding: 20px;
+	border-right: 1px solid #e0e0e0;
+	background-color: #f8f8f8;
+	margin-top: 60px;
+	height: calc(100vh - 60px);
+	overflow: auto;
+	z-index: 800;
+}
+
+.sidebarContainer {
+	display: flex;
+	flex-direction: column;
+	gap: 12px;
+}
+.paletteTitle {
+	margin-bottom: 20px;
+	font-size: 16px;
+	font-weight: 600;
+	color: #333;
+}
+
+.paletteNode {
+	padding: 14px;
+	color: #333;
+	border: 1px solid rgba(0, 0, 0, 0.06);
+	border-radius: 8px;
+	cursor: grab;
+	display: flex;
+	align-items: center;
+	font-size: 14px;
+	box-shadow: 0 2px 5px rgba(0, 0, 0, 0.06);
+	transition: transform 0.15s ease, box-shadow 0.15s ease;
+	font-weight: 500;
+}
+
+.deployButton {
+	padding: 12px;
+	background-color: #2684ff;
+	color: white;
+	border: none;
+	border-radius: 6px;
+	cursor: pointer;
+	font-weight: 500;
+	font-size: 14px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	gap: 8px;
+	transition: background-color 0.2s ease;
+	box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+}
+
+.deployButtonRunning {
+	background-color: #88b7d9;
+	cursor: not-allowed;
+}
+
+.clearButton {
+	padding: 12px;
+	background-color: #f8f8f8;
+	color: #666;
+	border: 1px solid #ddd;
+	border-radius: 6px;
+	cursor: pointer;
+	font-weight: 500;
+	font-size: 14px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	gap: 8px;
+	transition: all 0.2s ease;
+}
+
+.clearButtonHover {
+	background-color: #f3f3f3;
+	border-color: #ccc;
+}
+
+.configModalOverlay {
+	position: fixed;
+	top: 0;
+	left: 0;
+	right: 0;
+	bottom: 0;
+	background-color: rgba(0, 0, 0, 0.5);
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	z-index: 1100;
+}
+
+.configModal {
+	width: 500px;
+	background-color: white;
+	border-radius: 8px;
+	padding: 24px;
+	box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
+}
+
+.configModalTitle {
+	margin-top: 0;
+	margin-bottom: 20px;
+	font-size: 18px;
+	font-weight: 600;
+}
+
+.configModalLabel {
+	display: block;
+	margin-bottom: 8px;
+	font-weight: 500;
+	font-size: 14px;
+}
+
+.configModalInput {
+	width: 100%;
+	padding: 10px 12px;
+	border-radius: 6px;
+	border: 1px solid #ddd;
+	font-size: 14px;
+	box-sizing: border-box;
+}
+
+.configModalTextarea {
+	width: 100%;
+	height: 120px;
+	padding: 10px 12px;
+	border-radius: 6px;
+	border: 1px solid #ddd;
+	font-size: 14px;
+	box-sizing: border-box;
+	font-family: monospace;
+	resize: vertical;
+}
+
+.configModalActions {
+	display: flex;
+	justify-content: flex-end;
+	gap: 12px;
+}
+
+.configModalCancel {
+	padding: 10px 16px;
+	background-color: #f8f8f8;
+	border: 1px solid #ddd;
+	border-radius: 6px;
+	font-size: 14px;
+	cursor: pointer;
+}
+
+.configModalSave {
+	padding: 10px 16px;
+	background-color: #2684ff;
+	color: white;
+	border: none;
+	border-radius: 6px;
+	font-size: 14px;
+	cursor: pointer;
+}
+
+.reactflowWrapper {
+	flex: 1;
+	position: relative;
+	height: calc(100vh - 60px);
+	margin-top: 60px;
+}
+.reactflowControls {
+	border-radius: 8px;
+	box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+	background-color: #ffffff;
+}
+
+.reactflowMiniMap {
+	right: 12px !important;
+	bottom: 12px !important;
+	border-radius: 8px;
+	box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+}
